requirements.txt
flask
pathlib
python-dotenv
PyJWT
staffspy[browser]
scrapingbee